% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAAI.R
\name{calcValidAAI}
\alias{calcValidAAI}
\title{calcValidAAI}
\usage{
calcValidAAI(datasource = "LUH3")
}
\arguments{
\item{datasource}{Currently available: \code{"LUH3"}, \code{"LUH2v2"}, and \code{"GMIA"}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical area actually irrigated.
}
\author{
Stephen Wirth, Anne Biewald
}
