% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAgEmployment.R
\name{calcValidAgEmployment}
\alias{calcValidAgEmployment}
\title{calcValidAgEmployment}
\usage{
calcValidAgEmployment(datasource = "ILO")
}
\arguments{
\item{datasource}{So far only "ILO".}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
number of people employed in agriculture (crop+livestock production)
}
\examples{
\dontrun{
calcOutput("ValidAgEmployment", datasource="ILO")
}

}
\author{
Debbora Leip
}
