% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAgFFGDP.R
\name{calcValidAgFFGDP}
\alias{calcValidAgFFGDP}
\title{calcValidAgFFGDP}
\usage{
calcValidAgFFGDP(datasource = "WDI")
}
\arguments{
\item{datasource}{datasource for validation (WDI)}
}
\value{
List of magpie object with results on country level, no weight, unit and description.
}
\description{
validation for agricultural, fisheries, forestry value added gdp (Million 17USD)
}
\examples{
\dontrun{
calcOutput("ValidAgFFGDP")
}
}
\author{
David Chen
}
