% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAgGDP.R
\name{calcValidAgGDP}
\alias{calcValidAgGDP}
\title{calcValidAgGDP}
\usage{
calcValidAgGDP(datasource = "WDI")
}
\arguments{
\item{datasource}{datasource for validation, WDI and FAO}
}
\value{
List of magpie object with results on country level, no weight, unit and description.
}
\description{
validation for agricultural, fisheries, forestry value added gdp (Million 05USD)
}
\examples{
\dontrun{
calcOutput("ValidAgGDP")
}
}
\author{
David Chen, Edna J. Molina Bacca
}
