% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAnnualCarbonLTS.R
\name{calcValidAnnualCarbonLTS}
\alias{calcValidAnnualCarbonLTS}
\title{calcValidAnnualCarbonLTS}
\usage{
calcValidAnnualCarbonLTS(datasource = "Lauk_et_al")
}
\arguments{
\item{datasource}{Currently available \code{"Lauk_et_al"} or \code{"Johnston_Radeloff"} or \code{"Johnston_Radeloff_P"} (p = projection)}
}
\value{
List of magpie object with Emissions in wood products
}
\description{
Returns historical Emissions stored in wood products
}
\author{
Abhijeet Mishra
}
