% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidBEYield.R
\name{calcValidBEYield}
\alias{calcValidBEYield}
\title{calcValidBEYield}
\usage{
calcValidBEYield(cellular = FALSE)
}
\arguments{
\item{cellular}{determines whether data is returned at grid cell level
For aggregate = FALSE, cellular should be set to TRUE
For aggregate = "region+global", cellular should be set to FALSE
such that iso countries can be returned. Default is FALSE.}
}
\value{
List of magpie objects with results on cellular level, weight, unit, min, max, description.
}
\description{
Validation dataset for bioenergy crop yields from Li et al. (2020).
Each of the five crop types is returned as an independent entry at cellular level.
}
\examples{
\dontrun{
calcOutput("ValidBEYield", aggregate = FALSE)
}
}
\seealso{
[mrcommons::calcBEYield()]
}
\author{
Kristine Karstens, Felicitas Beier, Patrick Rein
}
