% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCarbonDensity.R
\name{calcValidCarbonDensity}
\alias{calcValidCarbonDensity}
\title{calcValidCarbonDensity}
\usage{
calcValidCarbonDensity(datasource = "LPJ_IPCC2006")
}
\arguments{
\item{datasource}{Datasources for validation data, e.g. LPJ_IPCC2006, LPJmL_natural}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for the carbon densities (including weights for aggregation)
}
\examples{

\dontrun{ 
  calcOutput("ValidCarbonDensity")
}

}
\seealso{
\code{\link{calcSOM}}
\code{\link{calcValidCarbonStocks}}
}
\author{
Kristine Karstens
}
