% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCarbonStocks.R
\name{calcValidCarbonStocks}
\alias{calcValidCarbonStocks}
\title{calcValidCarbonStocks}
\usage{
calcValidCarbonStocks(datasource = "LPJ_IPCC2006", baseyear = 1995)
}
\arguments{
\item{datasource}{Datasources for validation data, e.g. LPJ_IPCC2006, LPJmL_natural}

\item{baseyear}{baseyear for calculating soil carbon stock change}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for the carbon pools
}
\examples{

\dontrun{ 
  calcOutput("ValidCarbonStocks")
}

}
\seealso{
\code{\link{calcFoodSupplyPast}},
\code{\link{calcValidLivestockShare}}
}
\author{
Benjamin Leon Bodirsky
}
