% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCellularCroparea.R
\name{calcValidCellularCroparea}
\alias{calcValidCellularCroparea}
\title{calcValidCellularCroparea}
\usage{
calcValidCellularCroparea()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Croparea by crops on 0.5 degree grid in physical area
}
\examples{

\dontrun{ 
calcOutput("ValidCellularCroparea")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
