% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCellularCroplandNitrogenInputs.R
\name{calcValidCellularCroplandNitrogenInputs}
\alias{calcValidCellularCroplandNitrogenInputs}
\title{calcValidCellularCroplandNitrogenInputs}
\usage{
calcValidCellularCroplandNitrogenInputs()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nitrogen Inputs by crop type on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidCellularCroplandNitrogenInputs")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
