% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCellularLand.R
\name{calcValidCellularLand}
\alias{calcValidCellularLand}
\title{calcValidCellularLand}
\usage{
calcValidCellularLand()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports the main Land types on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidCellularLand")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
