% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCellularNitrogenBudgetCropland.R
\name{calcValidCellularNitrogenBudgetCropland}
\alias{calcValidCellularNitrogenBudgetCropland}
\title{calcValidCellularNitrogenBudgetCropland}
\usage{
calcValidCellularNitrogenBudgetCropland()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nitrogen Budget for Croplands on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidCellularNitrogenBudgetCropland")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
