% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCellularNitrogenBudgetPasture.R
\name{calcValidCellularNitrogenBudgetPasture}
\alias{calcValidCellularNitrogenBudgetPasture}
\title{calcValidCellularNitrogenBudgetPasture}
\usage{
calcValidCellularNitrogenBudgetPasture()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nitrogen Budgets for Pastures on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidCellularNitrogenBudgetPasture")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
