% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCellularYields.R
\name{calcValidCellularYields}
\alias{calcValidCellularYields}
\title{calcValidCellularYields}
\usage{
calcValidCellularYields()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Yields on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidCellularYields")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
