% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCMIP6.R
\name{calcValidCMIP6}
\alias{calcValidCMIP6}
\title{calcValidCMIP6}
\usage{
calcValidCMIP6()
}
\value{
GLO MAgPIE object containing average warming per RCP
}
\description{
validation for the MAGICC warming output, showing average temperature for the different RCPs
}
\examples{

\dontrun{
   calcOutput("calcValidCMIP6")
}

}
\author{
Michael Crawford
}
