% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidConsumptionValue.R
\name{calcValidConsumptionValue}
\alias{calcValidConsumptionValue}
\title{calcValidConsumptionValue}
\usage{
calcValidConsumptionValue(datasource = "FAO")
}
\arguments{
\item{datasource}{datasource for validation (FAO)}
}
\value{
List of magpie object with results on country level, no weight, unit and description.
}
\description{
Validation for consumption Value
}
\examples{
\dontrun{
calcOutput("ValidConsumptionValue")
}
}
\author{
Edna J. Molina Bacca
}
