% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCostsAEI.R
\name{calcValidCostsAEI}
\alias{calcValidCostsAEI}
\title{calcValidCostsAEI}
\usage{
calcValidCostsAEI(datasource = "IMPACT")
}
\arguments{
\item{datasource}{Datasource of validation data}
}
\value{
magpie object on country level, unit and description
}
\description{
calculates the validation data for irrigation investment costs
}
\examples{
\dontrun{ 
calcOutput("ValidCostsAEI")
}

}
\author{
Felicitas Beier
}
