% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCostsCapStocks.R
\name{calcValidCostsCapStocks}
\alias{calcValidCostsCapStocks}
\title{calcValidCostsCapStocks}
\usage{
calcValidCostsCapStocks(datasource = "FAO")
}
\arguments{
\item{datasource}{datasource for validation. FAO Value of Production dataset.}
}
\value{
List of magpie object with results on country level, no weight, unit and description.
}
\description{
Returns historical values for capital related costs
}
\examples{
\dontrun{
calcOutput("calcValidCostsCapStocks")
}
}
\author{
Edna J. Molina Bacca
}
