% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCostsFertilizer.R
\name{calcValidCostsFertilizer}
\alias{calcValidCostsFertilizer}
\title{calcValidCostsFertilizer}
\usage{
calcValidCostsFertilizer(datasource = "Vittis")
}
\arguments{
\item{datasource}{Datasource of validation data.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates validation data for fertilizer costs
}
\examples{

\dontrun{ 
calcOutput("ValidCostsFertilizer")
}

}
\author{
Debbora Leip
}
