% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCostsLabor.R
\name{calcValidCostsLabor}
\alias{calcValidCostsLabor}
\title{calcValidCostsLabor}
\usage{
calcValidCostsLabor(datasource = "Vittis")
}
\arguments{
\item{datasource}{Datasource of validation data.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates validation data for labor costs
}
\examples{

\dontrun{ 
calcOutput("ValidCostsLabor")
}

}
\author{
Debbora Leip
}
