% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCostsOverall.R
\name{calcValidCostsOverall}
\alias{calcValidCostsOverall}
\title{calcValidCostsOverall}
\usage{
calcValidCostsOverall(datasource = "FAO")
}
\arguments{
\item{datasource}{datasource for validation. FAO Value of Production dataset.}
}
\value{
List of magpie object with results on country level, no weight, unit and description.
}
\description{
Returns historical values of the overall value of production for the Agricultural,
Forestry sectors, and fisheries (mio.17USD)
}
\examples{
\dontrun{
calcOutput("calcValidCostsOverall")
}

}
\author{
Edna J. Molina Bacca
}
