% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCostsTransport.R
\name{calcValidCostsTransport}
\alias{calcValidCostsTransport}
\title{calcValidCostsTransport}
\usage{
calcValidCostsTransport(datasource = "GTAP")
}
\arguments{
\item{datasource}{Datasource of validation data.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for transport costs
}
\examples{
\dontrun{
calcOutput("ValidCostsTransport")
}

}
\author{
David Chen
}
