% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCropareaDiversity.R
\name{calcValidCropareaDiversity}
\alias{calcValidCropareaDiversity}
\title{calcValidCropareaDiversity}
\usage{
calcValidCropareaDiversity(index = "shannon", groupdiv = "agg1")
}
\arguments{
\item{index}{can be "shannon", "gini" or "invsimpson" for different types of diversitiy indices}

\item{groupdiv}{should crop groups be split up into several individual items or not?
Choose either FALSE or different (dis)aggregation methods "agg1", "agg2"}
}
\value{
MAgPIE object (unit depends on attributes)
}
\description{
calculates validation for croparea diversity index. As opposed to
CropareaDiversityIndex from magpie4 due to lack of data fallow land is not considered
}
\examples{
\dontrun{
x <- calcOutput("ValidCropareaDiversity", index = "shannon", groupdiv = "agg1", aggregate = FALSE)
}

}
\author{
Patrick v. Jeetze, Benjamin Leon Bodirsky
}
