% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCumulativeCarbonLTS.R
\name{calcValidCumulativeCarbonLTS}
\alias{calcValidCumulativeCarbonLTS}
\title{calcValidCumulativeCarbonLTS}
\usage{
calcValidCumulativeCarbonLTS(datasource = "Johnston_Radeloff")
}
\arguments{
\item{datasource}{Currently available \code{"Johnston_Radeloff"}}
}
\value{
List of magpie object with Emissions in wood products
}
\description{
Returns historical Emissions stored in wood products
}
\author{
Abhijeet Mishra
}
