% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidDemand.R
\name{calcValidDemand}
\alias{calcValidDemand}
\title{calcValidDemand}
\usage{
calcValidDemand(datasource = "FAO", detail = TRUE, nutrient = "dm")
}
\arguments{
\item{datasource}{Datasource of validation data.}

\item{detail}{if FALSE, only larger product categories are reported}

\item{nutrient}{The nutrient in which the results shall be reported.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for the utilization of agricultural products
for food, feed, bioenergy, seed, material, processing, or waste
}
\examples{

\dontrun{
calcOutput("ValidDemand")
}

}
\seealso{
\code{\link[mrcommons]{calcFoodSupplyPast}},
\code{\link{calcValidLivestockShare}}
}
\author{
Benjamin Leon Bodirsky, Isabelle Weindl
}
