% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidDemandBioenergy.R
\name{calcValidDemandBioenergy}
\alias{calcValidDemandBioenergy}
\title{calcValidDemandBioenergy}
\usage{
calcValidDemandBioenergy()
}
\value{
list of magpie object with data and weight
}
\description{
Returns future projections of bioenergy demand
}
\author{
Florian Humpenoeder
}
