% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidEmisLucGasser.R
\name{calcValidEmisLucGasser}
\alias{calcValidEmisLucGasser}
\title{calcValidEmisLucGasser}
\usage{
calcValidEmisLucGasser(subtype = "bookkeeping")
}
\arguments{
\item{subtype}{Available subtypes are:
\itemize{
\item historical:
\itemize{
\item Gasser_2020
\item LUH2_GCB_2019
\item FRA_2015
\item bookkeeping
}
}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical LUC emissions
}
\author{
Abhijeet Mishra
}
