% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidEmisLucGlo.R
\name{calcValidEmisLucGlo}
\alias{calcValidEmisLucGlo}
\title{calcValidEmisLucGlo}
\usage{
calcValidEmisLucGlo(subtype = "Canadell_2007")
}
\arguments{
\item{subtype}{Available subtypes are:
\itemize{
\item historical:
\itemize{
\item Canadell_2007
\item Friedlingstein_2010
\item Harris_2013
\item Houghton_2012
\item RCP
}
}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical and projected water usage from different sources
}
\author{
Florian Humpenoeder
}
