% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidEmisLucGloGasser.R
\name{calcValidEmisLucGloGasser}
\alias{calcValidEmisLucGloGasser}
\title{calcValidEmisLucGloGasser}
\usage{
calcValidEmisLucGloGasser(subtype = "Gasser_2020")
}
\arguments{
\item{subtype}{Available subtypes are:
\itemize{
\item historical:
\itemize{
\item Gasser_2020
\item LUH2_GCB_2019
\item FRA_2015
}
}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical LUC emissions
}
\author{
Abhijeet Mishra
}
