% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidEmissions.R
\name{calcValidEmissions}
\alias{calcValidEmissions}
\title{calcValidEmissions}
\usage{
calcValidEmissions(datasource = "CEDS")
}
\arguments{
\item{datasource}{The Emission Inventory that shall be used. For futher
information, best see mrcommons function calcEmissionInventory. Options are
e.g.  CEDS, combined_CEDS_IPCC (including own estimates where available),
IPCC(own estimates), Nsurplus (own estimates)}
}
\value{
List of magpie object with results on country level, weight on
country level, unit and description.
}
\description{
validation fo emissions
}
\examples{
\dontrun{
calcOutput("ValidEmissions")
}

}
\author{
Benjamin Leon Bodirsky
}
