% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidEmissionsAFOLU.R
\name{calcValidEmissionsAFOLU}
\alias{calcValidEmissionsAFOLU}
\title{ValidEmissionsAFOLU}
\usage{
calcValidEmissionsAFOLU(datasource = "FAO", cumulative = FALSE)
}
\arguments{
\item{datasource}{current options are "FAO" and "EDGAR_LU"}

\item{cumulative}{cumulative from y2000}
}
\value{
MAgPIE object with emissions in CO2eq
}
\description{
validation for total and cumulative AFOLU emissions in CO2eq
}
\examples{
\dontrun{
calcOutput("ValidEmissionsAFOLU")
}
}
\author{
Michael Crawford
}
