% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidEmissionsPeatland.R
\name{calcValidEmissionsPeatland}
\alias{calcValidEmissionsPeatland}
\title{calcValidEmissionsPeatland}
\usage{
calcValidEmissionsPeatland()
}
\value{
Global-level MAgPIE object with peatlands emissions in CO2eq for the year 2015
}
\description{
Validation data for peatland emissions in CO2eq, taken from the article:
     Leifeld, J., Menichetti, L. The underappreciated potential of peatlands in global
     climate change mitigation strategies. Nat Commun 9, 1071 (2018).
     https://doi.org/10.1038/s41467-018-03406-6
}
\examples{
\dontrun{
   calcOutput("ValidEmissionsPeatland")
}
}
\author{
Michael Crawford
}
