% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidFactorReqShares.R
\name{calcValidFactorReqShares}
\alias{calcValidFactorReqShares}
\title{calcValidFactorReqShares}
\usage{
calcValidFactorReqShares(subtype = "crop")
}
\arguments{
\item{subtype}{for which to report requirement shares, either "crop" or "livestock"}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for labor and capital requirement shares in agriculture
(currently the same shares for crop and livestock production based on USDA data)
}
\examples{

\dontrun{
calcOutput("ValidFactorReqShares")
}

}
\author{
Debbora Leip
}
