% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidFoodExpenditureShare.R
\name{calcValidFoodExpenditureShare}
\alias{calcValidFoodExpenditureShare}
\title{calcValidFoodExpenditureShare}
\usage{
calcValidFoodExpenditureShare(detail = FALSE, expenditureType = "agPrimary")
}
\arguments{
\item{detail}{if FALSE, only major food commoditiy groups are shown.}

\item{expenditureType}{Either "agPrimary" for agricultural primary products share,
or "food" for total food expenditure share including value-added marketing margins}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
validation for food expenditure share
}
\examples{
\dontrun{
calcOutput("ValidFoodExpenditureShare")
}

}
\author{
Benjamin Leon Bodirsky, David M Chen
}
