% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGini.R
\name{calcValidGini}
\alias{calcValidGini}
\title{calcValidGini}
\usage{
calcValidGini()
}
\value{
list of magpie object with data and weight.
}
\description{
Returns historical development of Gini Coefficient
}
\author{
David M Chen
}
