% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGlobalSurfaceTemp.R
\name{calcValidGlobalSurfaceTemp}
\alias{calcValidGlobalSurfaceTemp}
\title{calcValidGlobalSurfaceTemp}
\usage{
calcValidGlobalSurfaceTemp()
}
\value{
Global MAgPIE object
}
\description{
validation for the global surface temperature
}
\examples{

\dontrun{
   calcOutput("calcValidGlobalSurfaceTemp")
}

}
\author{
Michael Crawford
}
