% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGrassLSUha.R
\name{calcValidGrassLSUha}
\alias{calcValidGrassLSUha}
\title{calcValidGrassLSUha}
\usage{
calcValidGrassLSUha(datasource = "MAgPIEown")
}
\arguments{
\item{datasource}{Currently available: \code{"MAgPIEown"}}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for production of grass from managed pastures and rangelands separetely
}
\examples{
\dontrun{
calcOutput("ValidGrassLSUha")
}
}
\seealso{
\code{\link[mrcommons]{calcFAOmassbalance}},
}
\author{
Marcos Alves, Bin Lin
}
