% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGrassSoilCarbon.R
\name{calcValidGrassSoilCarbon}
\alias{calcValidGrassSoilCarbon}
\title{calcValidGrassSoilCarbon}
\usage{
calcValidGrassSoilCarbon(
  datasource = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1965-2100",
  model = "9eaf9b",
  lpjml
)
}
\arguments{
\item{datasource}{Datasources for validation data}

\item{model}{trained model ID}

\item{lpjml}{lpjml version}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for grasslands
}
\examples{
\dontrun{
calcOutput("ValidGrassSoilCarbon")
}
}
\author{
Marcos Alves
}
