% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridCarbonStocks.R
\name{calcValidGridCarbonStocks}
\alias{calcValidGridCarbonStocks}
\title{calcValidGridCarbonStocks}
\usage{
calcValidGridCarbonStocks(datasource = "LPJ_IPCC2006", baseyear = 1995)
}
\arguments{
\item{datasource}{Datasources for validation data, e.g. LPJ_IPCC2006, LPJmL_natural, ...}

\item{baseyear}{baseyear for calculating soil carbon stock change (for LPJ_IPCC2006 only)}
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
calculates the validation data for the carbon pools
}
\examples{

\dontrun{ 
  calcOutput("ValidGridCarbonStocks")
}

}
\author{
Kristine Karstens
}
