% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridCroplandNitrogenInputs.R
\name{calcValidGridCroplandNitrogenInputs}
\alias{calcValidGridCroplandNitrogenInputs}
\title{calcValidGridCroplandNitrogenInputs}
\usage{
calcValidGridCroplandNitrogenInputs()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nitrogen Inputs by crop type on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidGridCroplandNitrogenInputs")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
