% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridCroplandNitrogenSurplus.R
\name{calcValidGridCroplandNitrogenSurplus}
\alias{calcValidGridCroplandNitrogenSurplus}
\title{calcValidGrid}
\usage{
calcValidGridCroplandNitrogenSurplus()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Cropland Nitrogen Surplus and a Balanceflow that balances for unrealistically high nitrogen use efficiencies on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidGridCroplandNitrogenSurplus")
}

}
\seealso{
\code{\link{fullMADRATTOLPJML}}
}
\author{
Benjamin Leon Bodirsky
}
