% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridLand.R
\name{calcValidGridLand}
\alias{calcValidGridLand}
\title{calcValidGridLand}
\usage{
calcValidGridLand()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports the main Land types on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidGridLand")
}

}
\seealso{
\code{\link{fullMADRATTOLPJML}}
}
\author{
Benjamin Leon Bodirsky
}
