% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridNitrogenBudgetNonagland.R
\name{calcValidGridNitrogenBudgetNonagland}
\alias{calcValidGridNitrogenBudgetNonagland}
\title{calcValidGridNitrogenBudgetNonagland}
\usage{
calcValidGridNitrogenBudgetNonagland()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nitrogen Budget for non-agricultural land types on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidGridNitrogenBudgetNonagland")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
