% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridNitrogenBudgetPasture.R
\name{calcValidGridNitrogenBudgetPasture}
\alias{calcValidGridNitrogenBudgetPasture}
\title{calcValidGridNitrogenBudgetPasture}
\usage{
calcValidGridNitrogenBudgetPasture()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nitrogen Budgets for Pastures on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidGridNitrogenBudgetPasture")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
