% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridNutrientAWMS.R
\name{calcValidGridNutrientAWMS}
\alias{calcValidGridNutrientAWMS}
\title{calcValidGridNutrientAWMS}
\usage{
calcValidGridNutrientAWMS(nutrient = c("nr", "c"))
}
\arguments{
\item{nutrient}{can be c, nr, p, k. For p and k, no losses are assumed in confinements.}
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nutrient load in animal waste management on 0.5 degree grid
}
\examples{
\dontrun{
calcOutput("ValidGridNitrogenBudgetCropland")
}

}
\seealso{
\code{\link{fullMADRATTOLPJML}}
}
\author{
Kristine Karstens, Benjamin Leon Bodirsky
}
