% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridNutrientLossesAWMS.R
\name{calcValidGridNutrientLossesAWMS}
\alias{calcValidGridNutrientLossesAWMS}
\title{calcValidGridNutrientLossesAWMS}
\usage{
calcValidGridNutrientLossesAWMS(nutrient = c("nr", "c"))
}
\arguments{
\item{nutrient}{can be c, nr, p, k. For p and k, no losses are assumed in confinements.}
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Nutrient Losses in animal waste management on 0.5 degree grid
}
\examples{
\dontrun{
calcOutput("ValidGridNitrogenBudgetCropland")
}

}
\author{
Benjamin Leon Bodirsky
}
