% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridResidueDemand.R
\name{calcValidGridResidueDemand}
\alias{calcValidGridResidueDemand}
\title{calcValidGridResidueDemand}
\usage{
calcValidGridResidueDemand()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports aboveground Crop Residue Demand on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidGridResidueDemand")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
