% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridSOCStocks.R
\name{calcValidGridSOCStocks}
\alias{calcValidGridSOCStocks}
\title{calcValidGridSOCStocks}
\usage{
calcValidGridSOCStocks(
  datasource = "LPJ_IPCC2006",
  baseyear = 1995,
  intensive = FALSE
)
}
\arguments{
\item{datasource}{Datasources for validation data, e.g. LPJ_IPCC2006, SoilGrids, ...}

\item{baseyear}{baseyear for calculating soil carbon stock change (for LPJ_IPCC2006 only)}

\item{intensive}{If FALSE (default) total stocks will be returned; otherwise (TRUE) carbon densities.}
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
calculates the validation data for the gridded soil carbon pools
}
\examples{
\dontrun{
calcOutput("ValidGridSOCStocks")
}

}
\author{
Kristine Karstens
}
