% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridYields.R
\name{calcValidGridYields}
\alias{calcValidGridYields}
\title{calcValidGridYields}
\usage{
calcValidGridYields(
  datasource = "downscaledFAO",
  future = NULL,
  physical = TRUE
)
}
\arguments{
\item{datasource}{downscaledFAO or calibratedFAO}

\item{future}{if NULL no future values are returned (default).
specify climate scenario (gcm:rcp), if future is needed}

\item{physical}{if true (default) physical area (croparea) used for yield calculation;
if false harvested area used for yield calculation}
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Yields on 0.5 degree grid
}
\examples{
\dontrun{
calcOutput("ValidGridYields")
}

}
\seealso{
\code{\link{fullMADRATTOLPJML}}
}
\author{
Benjamin Leon Bodirsky
}
