% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGrowingStock.R
\name{calcValidGrowingStock}
\alias{calcValidGrowingStock}
\title{calcValidGrowingStock}
\usage{
calcValidGrowingStock(datasource = "FAO")
}
\arguments{
\item{datasource}{Currently only  available for the "FAO" source}
}
\value{
List of magpie object with growing stock
}
\description{
Returns historical FRA 2015 growing stock in mio m3 per ha
}
\author{
Abhijeet Mishra
}
