% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidKcal.R
\name{calcValidKcal}
\alias{calcValidKcal}
\title{calcValidKcal}
\usage{
calcValidKcal(datasource = "FAO", nutrient = "kcal", detail = TRUE)
}
\arguments{
\item{datasource}{Datasource of validation data. If "FAO", we use FAO calories with FAO population data (slightly diverges from original data as the convert script for example splits up countries for the past). In the case of "PopulationPast", we also use FAO calorie values, but divide them by our standard population. "Bodirsky2015" uses projections from slightly altered method of the Bodirsky et al 2015 PLOS ONE paper.}

\item{nutrient}{kcal or protein}

\item{detail}{if FALSE, only larger product categories are reported}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for calorie food supply
}
\examples{

\dontrun{ 
calcOutput("ValidKcal")
}

}
\seealso{
\code{\link{calcFoodSupplyPast}},
\code{\link{calcValidLivestockShare}}
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens
}
