% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidLand.R
\name{calcValidLand}
\alias{calcValidLand}
\title{calcValidLand}
\usage{
calcValidLand(datasource = "MAgPIEown")
}
\arguments{
\item{datasource}{Currently available: \code{"FAO"}, \code{"LUH2v2"}, \code{"MAgPIEown"} and \code{"SSPResults"}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical cropland, pasture and forest area from FAOSTAT that can
be used for model validation.
}
\author{
Ulrich Kreidenweis, Benjamin Bodirsky, Abhijeet Mishra, Mishko Stevanovic, Kristine Karstens
}
