% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidLandChange.R
\name{calcValidLandChange}
\alias{calcValidLandChange}
\title{calcValidLandChange}
\usage{
calcValidLandChange(baseyear = 1995, datasource = "MAgPIEown")
}
\arguments{
\item{baseyear}{baseyear for calculating land-use change}

\item{datasource}{Currently available: \code{"FAO"}, \code{"LUH2v2"}, \code{"LUH3"}, \code{"MAgPIEown"}
and \code{"SSPResults"}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical changes of cropland, pasture and forest area from LUH2 and FAOSTAT that can
be used for model validation.
}
\author{
Florian Humpenoeder
}
