% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidLivestockShare.R
\name{calcValidLivestockShare}
\alias{calcValidLivestockShare}
\title{calcValidLivestockShare}
\usage{
calcValidLivestockShare(datasource = "FAO")
}
\arguments{
\item{datasource}{Datasource of validation data. If "FAO", we use FAO calories with FAO population data (slightly diverges from original data as the convert script for example splits up countries for the past). In the case of "PopulationPast", we also use FAO calorie values, but divide them by our standard population}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for the share of livestock products (including fish) in total calorie food supply
}
\examples{

\dontrun{
calcOutput("ValidLivestockShare")
}

}
\seealso{
\code{\link[mrcommons]{calcFoodSupplyPast}},
\code{\link{calcValidKcal}}
}
\author{
Benjamin Leon Bodirsky
}
