% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidManure.R
\name{calcValidManure}
\alias{calcValidManure}
\title{calcValidManure}
\usage{
calcValidManure(datasource = "Bodirsky")
}
\arguments{
\item{datasource}{own: own estimation for the past based on feed intake, IPCC: standard excretion factors, FAO: FAO estimates}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Validates the esitmates on  excretion and manure management
}
\examples{

\dontrun{
calcOutput("ValidManure")
}

}
\seealso{
\code{\link{calcExcretion}},
\code{\link{calcExcretionIPCC}}
}
\author{
Benjamin Leon Bodirsky
}
