% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidMapBiomaslu.R
\name{calcValidMapBiomaslu}
\alias{calcValidMapBiomaslu}
\title{calcValidMapBiomaslu}
\usage{
calcValidMapBiomaslu()
}
\value{
List of magpie objects with results for Brazil, unit and description.
}
\description{
Integrates the MapBiomas landuse-dataset.
}
\examples{

\dontrun{ 
calcOutput("ValidMapBiomaslu")
}

}
\author{
Geanderson Ambrosio
}
