% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidNitrogenBudgetNonagland.R
\name{calcValidNitrogenBudgetNonagland}
\alias{calcValidNitrogenBudgetNonagland}
\title{calcValidNitrogenBudgetNonagland}
\usage{
calcValidNitrogenBudgetNonagland()
}
\value{
MAgPIE object with results on the iso level, unweighted, unit, and description
}
\description{
reports nitrogen budget for non-agricultural land types
}
\examples{

\dontrun{
calcOutput("calcValidNitrogenBudgetNonagland")
}
}
\author{
Michael Crawford
}
