% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidNitrogenBudgetPasture.R
\name{calcValidNitrogenBudgetPasture}
\alias{calcValidNitrogenBudgetPasture}
\title{calcValidNitrogenBudgetPasture}
\usage{
calcValidNitrogenBudgetPasture(datasource = "Bodirsky")
}
\arguments{
\item{datasource}{Bodirsky for own calculations, Lassaletta2014 for a country dataset from 
Lassaletta, L., G. Billen, B. Grizzetti, J. Angalde, and J. Garnier. 2014. 
50 Year Trends in Nitrogen Use Efficiency of World Cropping Systems: The Relationship between Yield and Nitrogen Input to Pasture.
Environmental Research Letters.
FAO for some N related parameters published in FAOSTAT.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Validation Script for Nitrogen Budgets on Pastures
}
\examples{

\dontrun{ 
calcOutput("ValidNitrogenBudgetPasture")
}

}
\seealso{
\code{\link[mrcommons]{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
