% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidNitrogenSurplus.R
\name{calcValidNitrogenSurplus}
\alias{calcValidNitrogenSurplus}
\title{calcValidNitrogenSurplus}
\usage{
calcValidNitrogenSurplus()
}
\value{
MAgPIE object with results on the iso level, unweighted, unit, and description
}
\description{
reports nutrient surplus for all land types
}
\examples{

\dontrun{
calcOutput("calcValidNitrogenSurplus")
}
}
\author{
Michael Crawford
}
