% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPlanetaryBoundariesPhosphorus.R
\name{calcValidPlanetaryBoundariesPhosphorus}
\alias{calcValidPlanetaryBoundariesPhosphorus}
\title{calcValidPlanetaryBoundariesPhosphorus}
\usage{
calcValidPlanetaryBoundariesPhosphorus(datasource = "Rockstroem2009")
}
\arguments{
\item{datasource}{datasource to compare to. Historical trajectories of the indicators, or the planetary boundary by the Rockstroem Paper 2009 or the Steffen et al paper 2015.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Provides comparison for the planetary boundary indicators for phosphorus
}
\examples{

\dontrun{ 
calcOutput("ValidPlanetaryBoundariesPhosphorus",datasource="Rockstroem2009",aggregate=FALSE)
}

}
\seealso{
\code{\link{calcValidPlanetaryBoundariesNitrogen}},
\code{\link{calcOutput}}
}
\author{
Benjamin Leon Bodirsky
}
