% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPopulation.R
\name{calcValidPopulation}
\alias{calcValidPopulation}
\title{calcValidPopulation}
\usage{
calcValidPopulation(
  pastData = "WDI-UN_PopDiv-MI",
  futureData = "SSPs-UN_PopDiv"
)
}
\arguments{
\item{pastData}{population past data source}

\item{futureData}{population future data source}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical development of population and future projections of population dynamics
}
\author{
Florian Humpenoeder, Jan Philipp Dietrich, David Chen
}
