% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPovertyLine.R
\name{calcValidPovertyLine}
\alias{calcValidPovertyLine}
\title{calcValidPovertyLine}
\usage{
calcValidPovertyLine(datasource = "WBPoverty", subtype = "320PovertyLine")
}
\arguments{
\item{datasource}{datasource for validation. Option: "WBPoverty" (world bank).}

\item{subtype}{Poverty line criteria. Options:"320PovertyLine" (default), "550PovertyLine", "1p90 USDppp11/day"}
}
\value{
List of magpie object with results on country level, no weight, unit and description.
}
\description{
Validation of poverty (head-count)
}
\examples{
\dontrun{
calcOutput("ValidPovertyLine")
}

}
\author{
Edna J. Molina Bacca
}
