% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPriceAgriculture.R
\name{calcValidPriceAgriculture}
\alias{calcValidPriceAgriculture}
\title{calcValidPriceAgriculture}
\usage{
calcValidPriceAgriculture(datasource = "FAO")
}
\arguments{
\item{datasource}{Options of the source of data:  \code{IMPACT3.2.2World_Price}, \code{FAO}, \code{FAOp} and \code{WBGEM}.}
}
\value{
List with a magpie object with commodity prices on global and country level.
}
\description{
provides global prices from the IMPACT model projections, World Bank observations, and FAO 
obersvations for MAgPIE commodities in $2017/tDM
}
\examples{

\dontrun{ 
calcOutput("ValidPriceAgriculture", datasource="IMPACT3.2.2World_Price", aggregate=FALSE)
calcOutput("ValidPriceAgriculture", datasource="FAO")
}

}
\seealso{
\code{\link{readIMPACT3.2.2World_Price}}
}
\author{
Mishko Stevanovic
}
