% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPriceBioenergy.R
\name{calcValidPriceBioenergy}
\alias{calcValidPriceBioenergy}
\title{calcValidPriceBioenergy}
\usage{
calcValidPriceBioenergy()
}
\value{
list of magpie object with data and weight
}
\description{
Returns future projections of biomass prices
}
\author{
Florian Humpenoeder
}
