% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPriceGHG.R
\name{calcValidPriceGHG}
\alias{calcValidPriceGHG}
\title{calcValidPriceGHG}
\usage{
calcValidPriceGHG(datasource = "SSPResults")
}
\arguments{
\item{datasource}{Currently available: \code{"SSPResults"}}
}
\description{
Validates MAgPIE GHG emission price input against SSP GHG emission price projections
}
\examples{
\dontrun{
calcOutput("ValidPriceGHG")
}

}
\author{
Amsalu W. Yalew, Benjamin Leon Bodirsky, Florian Humpenoeder
}
