% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPriceIndex.R
\name{calcValidPriceIndex}
\alias{calcValidPriceIndex}
\title{calcValidPriceIndex}
\usage{
calcValidPriceIndex(
  datasource = "FAO",
  value = "real",
  baseyear = "y2005",
  round = TRUE
)
}
\arguments{
\item{datasource}{Options of the source of data:  \code{FAO}.}

\item{value}{\code{real} and \code{nominal}}

\item{baseyear}{Baseyear for normalizing of price index.}

\item{round}{Rounding of price index to intiger numbers. Default \code{TRUE}.}
}
\value{
List with a magpie object with global nominal and real price index.
}
\description{
provides global producer price index from FAO
}
\examples{
\dontrun{
a <- calcOutput("ValidPriceIndex", value = "real", aggregate = FALSE)
}

}
\seealso{
\code{\link{readProdPrIndex}}
}
\author{
Mishko Stevanovic
}
