% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidProcessing.R
\name{calcValidProcessing}
\alias{calcValidProcessing}
\title{calcValidProcessing}
\usage{
calcValidProcessing(
  datasource = "FAO",
  detail = TRUE,
  nutrient = "dm",
  indicator = "primary_to_process"
)
}
\arguments{
\item{datasource}{Datasource of validation data.}

\item{detail}{if FALSE, only larger product categories are reported}

\item{nutrient}{The nutrient in which the results shall be reported.}

\item{indicator}{if "primary_to_process", returns the amount of primary product into processing.
if "secondary_from_primary", gives primary into secondary.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for the share of inputs that go into processing processes,
and the share of inputs that secondary products are composed of
}
\examples{
\dontrun{
calcOutput("ValidProcessing", indicator = "secondary_from_primary")
}

}
\seealso{
\code{\link[mrcommons]{calcFoodSupplyPast}},
\code{\link{calcValidLivestockShare}}
}
\author{
David Chen, Benjamin Leon Bodirsky
}
