% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidProduction.R
\name{calcValidProduction}
\alias{calcValidProduction}
\title{calcValidProduction}
\usage{
calcValidProduction(datasource = "FAO", detail = TRUE, nutrient = "dm")
}
\arguments{
\item{datasource}{Datasource of validation data.}

\item{detail}{if FALSE, only larger product categories are reported}

\item{nutrient}{The nutrient in which the results shall be reported.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for production of agricultural products
}
\examples{
\dontrun{
calcOutput("ValidProduction")
}

}
\seealso{
\code{\link{calcFAOmassbalance}},
\code{\link{calcValidDemand}}
}
\author{
Benjamin Leon Bodirsky
}
