% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidProductionGrowth.R
\name{calcValidProductionGrowth}
\alias{calcValidProductionGrowth}
\title{calcValidProductionGrowth}
\usage{
calcValidProductionGrowth(
  datasource = "FAO",
  detail = TRUE,
  base_year = "y1995"
)
}
\arguments{
\item{datasource}{which data to use}

\item{detail}{if true, provides results for all commodities, otherwhise aggregates some groups}

\item{base_year}{y1995}
}
\value{
Production growth rates (index)
}
\description{
calculate historical production growth rate
}
\examples{

  \dontrun{
    x <- calcValidProductionGrowth(datasource = "FAO", detail = TRUE,base_year ="y1995")
  }
}
\author{
Xiaoxi Wang
}
