% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidResidues.R
\name{calcValidResidues}
\alias{calcValidResidues}
\title{calcValidResidues}
\usage{
calcValidResidues()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for residues (biomass, field balance, demands)
}
\examples{
\dontrun{
calcOutput("ValidResidues")
}

}
\author{
Kristine Karstens
}
