% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidRotationLength.R
\name{calcValidRotationLength}
\alias{calcValidRotationLength}
\title{calcValidRotationLength}
\usage{
calcValidRotationLength(datasource = "FAO 2006")
}
\arguments{
\item{datasource}{Datasources for validation data}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for MAgPIE rotation lengths in plantations
}
\examples{
\dontrun{
calcOutput("ValidRotationLength")
}

}
\author{
Abhijeet Mishra
}
