% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidSDG1.R
\name{calcValidSDG1}
\alias{calcValidSDG1}
\title{calcValidSDG1}
\usage{
calcValidSDG1(datasource = "James")
}
\arguments{
\item{datasource}{Currently only  available for the "James" source}
}
\value{
List of magpie object with data and population
}
\description{
Returns historical SDG1 INDICATOR_POVERTY  IN USD05/cap/yr
}
\author{
Edna J. Molina Bacca
}
