% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidSDG12.R
\name{calcValidSDG12}
\alias{calcValidSDG12}
\title{calcValidSDG12}
\usage{
calcValidSDG12(datasource = "FAO")
}
\arguments{
\item{datasource}{FAO}
}
\value{
List of magpie object with data and population
}
\description{
Returns historical SDG12 Indicators_Sustainable Production and Consumption
}
\author{
Edna J. Molina Bacca
}
