% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidSelfsuff.R
\name{calcValidSelfsuff}
\alias{calcValidSelfsuff}
\title{calcValidSelfsuff}
\usage{
calcValidSelfsuff(datasource = "FAO", detail = TRUE)
}
\arguments{
\item{datasource}{Options of the source of data:  \code{FAO, FAOpre2010, FAOpost2010}.}

\item{detail}{Default is \code{TRU}. If \code{FALSE}, the subcategories of
groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Validates self-sufficiency ration
}
\examples{
\dontrun{
calcOutput("ValidSelfsuff")
}

}
\author{
Benjamin Leon Bodirsky, Mishko Stevanovic
}
