% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidSOCShare.R
\name{calcValidSOCShare}
\alias{calcValidSOCShare}
\title{calcValidSOCShare}
\usage{
calcValidSOCShare(datasource = "histSOCbudget")
}
\arguments{
\item{datasource}{Datasources only "histSOCbudget"}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for the soil carbon shares
}
\examples{
\dontrun{
calcOutput("ValidSOCShare")
}

}
\author{
Kristine Karstens
}
