% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidSOCStocks.R
\name{calcValidSOCStocks}
\alias{calcValidSOCStocks}
\title{calcValidSOCStocks}
\usage{
calcValidSOCStocks(datasource = "histSOCbudget", baseyear = 1995)
}
\arguments{
\item{datasource}{Datasources for validation data, e.g. LPJ_IPCC2006, SoilGrids}

\item{baseyear}{baseyear for calculating soil carbon stock change}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for the soil carbon pools
}
\examples{
\dontrun{
calcOutput("ValidSOCStocks")
}

}
\seealso{
\code{\link[mrcommons]{calcFoodSupplyPast}},
\code{\link{calcValidLivestockShare}}
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens
}
