% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidSSPResults.R
\name{calcValidSSPResults}
\alias{calcValidSSPResults}
\title{calcValidSSPResults}
\usage{
calcValidSSPResults()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Convert SSPResults to reporting format
}
\examples{

\dontrun{ 
  calcOutput("ValidSSPResults")
}

}
\seealso{
\code{\link{convertSSPResults}}
}
\author{
Kristine Karstens
}
