% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTauPastr.R
\name{calcValidTauPastr}
\alias{calcValidTauPastr}
\title{ValidTauPastr}
\usage{
calcValidTauPastr()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates managed pastures Tau based on FAO yield trends for 1995.
}
\examples{
\dontrun{
calcOutput("ValidTauPastr")
}

}
\author{
Marcos Alves
}
