% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTimberDemand.R
\name{calcValidTimberDemand}
\alias{calcValidTimberDemand}
\title{calcValidTimberDemand}
\usage{
calcValidTimberDemand(datasource = "FAO")
}
\arguments{
\item{datasource}{Currently only  available for the "FAO" source}
}
\value{
List of magpie object with data and population
}
\description{
Returns historical timber demand in mio m3 per yr
}
\author{
Abhijeet Mishra
}
