% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTradeGrowth.R
\name{calcValidTradeGrowth}
\alias{calcValidTradeGrowth}
\title{calcValidTradeGrowth}
\usage{
calcValidTradeGrowth(datasource = "FAO", detail = TRUE, base_year = "y1995")
}
\arguments{
\item{datasource}{which data to use}

\item{detail}{if true, provides results for all commodities, otherwhise aggregates some groups}

\item{base_year}{y1995}
}
\value{
Trade growth rates (index)
}
\description{
calculate historical trade growth rate
}
\examples{

  \dontrun{
    x <- calcValidTradeGrowth(datasource = "FAO", detail = TRUE,base_year ="y1995")
  }
}
\author{
Xiaoxi Wang
}
