% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidWageDevelopment.R
\name{calcValidWageDevelopment}
\alias{calcValidWageDevelopment}
\title{calcValidWageDevelopment}
\usage{
calcValidWageDevelopment(datasource = "ILO_completed")
}
\arguments{
\item{datasource}{So far only "ILO_completed" (based on regression between ILO hourly labor costs and GDP pc MER,
calibrated to match USDA/FAO total labor costs)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
wage index: hourly labor costs in crop+livestock production relative to 2000
}
\examples{
\dontrun{
calcOutput("ValidWageDevelopment", datasource = "ILO_completed")
}

}
\author{
Debbora Leip
}
