% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidWaterUsage.R
\name{calcValidWaterUsage}
\alias{calcValidWaterUsage}
\title{calcValidWaterUsage}
\usage{
calcValidWaterUsage(datasource = "shiklomanov_2000")
}
\arguments{
\item{datasource}{Currently available: 
\itemize{
\item historical:
\itemize{
\item \code{"foley_2011"}
\item \code{"shiklomanov_2000"}
\item \code{"wada_2011"}
\item \code{"wisser_2008"}
}
\item projections:
\itemize{
\item \code{"fischer_IIASA"}
\item \code{"hejazi_2013"}
 \item \code{molden_IWMI}
 \item \code{seckler_IWMI}
 \item \code{shiklomanov}
}
}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical and projected water usage from different sources
}
\author{
Stephen Wirth, Anne Biewald
}
