% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidYield.R
\name{calcValidYield}
\alias{calcValidYield}
\title{calcValidYield}
\usage{
calcValidYield(
  datasource = "MadratLandInGFAOLUH",
  faoVersion = "join2010",
  future = NULL,
  physical = TRUE
)
}
\arguments{
\item{datasource}{Specify which datasource needs to be used.
Currently only "MadratLandInGFAOLUH" and "calibratedLPJmL" is available.}

\item{faoVersion}{which FAO version to use (pre, post or joined 2010 data)}

\item{future}{if NULL no future values are returned (default).
specify climate scenario (gcm:rcp), if future is needed}

\item{physical}{if true (default) physical area (croparea) used for yield calculation;
if false harvested area used for yield calculation}
}
\value{
List of magpie objects with results on country level,
        weight on country level, unit,
        Max.&Min. values alongwith description.
}
\description{
Calculates a dataset of agricultural production out of the combined
data of calcFAOharmonized(). Covers dry matter (DM) production.
Also returns areas of individual crops from FAOSTAT.
Total area can be lower or higher than arable land because of multicropping or fallow land.
Returns yield as calculated from area area and production.
}
\examples{
\dontrun{
calcOutput("ValidYield")
}
}
\seealso{
\code{\link[mrcommons]{calcFAOmassbalance}},
\code{\link[mrlandcore]{calcCroparea}}
}
\author{
Abhijeet Mishra, Isabelle Weindl, Florian Humpenoeder
}
