% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGasser.R
\name{convertGasser}
\alias{convertGasser}
\title{convertGasser}
\usage{
convertGasser(x)
}
\arguments{
\item{x}{MAgPIE object containing Gasser emissions data}
}
\value{
Historical LUC emission data
}
\description{
Convert subtypes on ISO country level.
}
\details{
Convert Gasser Data
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Abhijeet Mishra
}
