% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGCBcountry.R
\name{convertGCBcountry}
\alias{convertGCBcountry}
\title{convertGCBcountry}
\usage{
convertGCBcountry(x)
}
\arguments{
\item{x}{magpie object returned by readGCBcountry}
}
\value{
magpie object on the 249 ISO countries
}
\description{
Fill the country-level GCB bookkeeping fluxes to the standard ISO country list.
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Florian Humpenoeder
}
