% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIMPACTIrrigInvCosts.R
\name{convertIMPACTIrrigInvCosts}
\alias{convertIMPACTIrrigInvCosts}
\title{convertIMPACTIrrigInvCosts}
\usage{
convertIMPACTIrrigInvCosts(x)
}
\arguments{
\item{x}{MAgPIE object containing irrigation investment cost data on region level}
}
\value{
magpie object containing average annual baseline water-related
        investment cost on country-level
        (in million 2017 USD per year)
}
\description{
converts units and dimensions of average annual baseline
             water-related investment cost data from readIMPACTIrrigInvCosts
}
\examples{
\dontrun{
a <- readSource("IMPACTIrrigInvCosts", convert = TRUE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Felicitas Beier
}
