% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPardeyAgRD.R
\name{convertPardeyAgRD}
\alias{convertPardeyAgRD}
\title{convertPardeyAgRD}
\usage{
convertPardeyAgRD(x)
}
\arguments{
\item{x}{MAgPIE object containing PardeyRD data}
}
\value{
magpie object containing expenditure in Ag R&D, 2009 USD PPP
}
\description{
Agricultural R&D investment data read from https://www.nature.com/news/agricultural-rd-is-on-the-move-1.20571
3 tables are read in: AgRD_Pardey is public Ag expenditure in 1960 and 2011, extracted from the interactive figure in the article that has more complete countries
agGERD and agPERD are total and public expenditures respectively, for less countries but more years
}
\examples{

\dontrun{
a <- convertSource("PardeyAgRD")
}
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
David Chen
}
