% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWaterUsage.R
\name{convertWaterUsage}
\alias{convertWaterUsage}
\title{convertWaterUsage}
\usage{
convertWaterUsage(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing IrrigatedArea data on Country level}

\item{subtype}{: No subtype needed}
}
\value{
Data on water use as MAgPIE object on country level Missing values are added as NA
}
\description{
Convert data on agricultural water use
Convert subtypes on ISO country level.
}
\examples{
\dontrun{
a <- readSource("WaterUsage", "aquastat_2008_12")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
