% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSoilCarbonDebt.R
\name{correctSoilCarbonDebt}
\alias{correctSoilCarbonDebt}
\title{correctSoilCarbonDebt}
\usage{
correctSoilCarbonDebt(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects
}
\description{
Correct data from Soil Carbon Debt Paper (https://github.com/whrc/Soil-Carbon-Debt/)
}
\examples{

\dontrun{
  readSource("SoilCarbonDebt", convert="onlycorrect")
}

}
\author{
Kristine Karstens
}
