% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAR6SPMFig1.R
\name{downloadAR6SPMFig1}
\alias{downloadAR6SPMFig1}
\title{downloadAR6SPMFig1}
\usage{
downloadAR6SPMFig1()
}
\value{
Metadata on downloaded dataset
}
\description{
download IPCC AR6 Summary for Policymakers Figure 1, global surface temperature change
relative to 1850-1900 from observations and CMIP6 models (human and natural forcings simulations)
}
\examples{

\dontrun{
  downloadSource("AR6SPMFig1")
}

}
\seealso{
https://www.ipcc.ch/report/ar6/wg1/downloads/report/IPCC_AR6_WGI_SPM.pdf
}
\author{
Michael Crawford
}
