% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCMIP6.R
\name{downloadCMIP6}
\alias{downloadCMIP6}
\title{downloadCMIP6}
\usage{
downloadCMIP6()
}
\value{
Metadata on downloaded CMIP near-surface air temperature data
}
\description{
Download near-surface air temperature projection from CMIP6 for the following scenarios:
SSP1-1.9, SSP1-2.6, SSP2-4.5, SSP3-7.0, SSP4-3.4, SSP4-6.0, SSP5-8.5, SSP5-3.4-over
}
\examples{

\dontrun{
  downloadSource("CMIP6")
}

}
\seealso{
https://cmip6.science.unimelb.edu.au/
}
\author{
Michael Crawford
}
