% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGCBcountry.R
\name{downloadGCBcountry}
\alias{downloadGCBcountry}
\title{downloadGCBcountry}
\usage{
downloadGCBcountry()
}
\description{
Download the Obermeier et al. (2024) country-level LULUCF carbon flux dataset
(net and gross fluxes of the GCB bookkeeping models), archived on Zenodo (8144174).
}
\examples{
\dontrun{
downloadSource("GCBcountry")
}
}
\seealso{
\code{\link[madrat]{downloadSource}}
}
\author{
Florian Humpenoeder
}
