% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGISTEMP.R
\name{downloadGISTEMP}
\alias{downloadGISTEMP}
\title{downloadGISTEMP}
\usage{
downloadGISTEMP()
}
\value{
Metadata on downloaded dataset
}
\description{
download GISTEMP historical global surface temperature data from the Goddard Institute for Space Studies
}
\examples{

\dontrun{
  downloadSource("GISTEMP")
}

}
\seealso{
https://data.giss.nasa.gov/gistemp/
}
\author{
Michael Crawford
}
