% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGlobalCarbonBudget.R
\name{downloadGlobalCarbonBudget}
\alias{downloadGlobalCarbonBudget}
\title{downloadGlobalCarbonBudget}
\usage{
downloadGlobalCarbonBudget()
}
\value{
Metadata from the Global Carbon Budget 2023 dataset
}
\description{
download the most current Global Carbon Budget dataset
}
\examples{
\dontrun{
downloadSource("GlobalCarbonBudget")
}

}
\author{
Michael Crawford
}
