% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSoilCarbonDebt.R
\name{downloadSoilCarbonDebt}
\alias{downloadSoilCarbonDebt}
\title{downloadSanderman}
\usage{
downloadSoilCarbonDebt()
}
\value{
Meta information on downloaded data
}
\description{
Download data from Soil Carbon Debt Paper (https://github.com/whrc/Soil-Carbon-Debt/)
}
\examples{

\dontrun{
  readSource("SoilCarbonDebt", subtype = "SOCS_2010")
}

}
\author{
Kristine Karstens
}
