% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullENGAGE.R
\name{fullENGAGE}
\alias{fullENGAGE}
\title{fullENGAGE}
\usage{
fullENGAGE()
}
\value{
validation data in IAMC reporting format
}
\description{
retrieves validation data from mrcommons for the ENGAGE project
}
\author{
Florian Humpenoeder
}
