% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMADRATtoLPJml.R
\name{fullMADRATTOLPJML}
\alias{fullMADRATTOLPJML}
\title{fullMADRATTOLPJML}
\usage{
fullMADRATTOLPJML()
}
\value{
No return; the function writes a number of netcdf files into the output folder of MADRAT
}
\description{
creates historical outputs to be used by LPJml or other models that require gridded inputs
}
\examples{

\dontrun{ 
fullMADRATTOLPJML()
}

}
\seealso{
\code{\link{fullVALIDATION}}
}
\author{
Benjamin Leon Bodirsky
}
