% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrvalidation-package.R
\docType{package}
\name{mrvalidation-package}
\alias{mrvalidation}
\alias{mrvalidation-package}
\title{mrvalidation: madrat data preparation for validation purposes}
\description{
Package contains routines to prepare data for validation exercises.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrvalidation}
  \item \url{https://doi.org/10.5281/zenodo.4317826}
  \item Report bugs at \url{https://github.com/pik-piam/mrvalidation/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}

Authors:
\itemize{
  \item Stephen Wirth
  \item Kristine Karstens
  \item Florian Humpenoeder
  \item Mishko Stevanovic
  \item Abhijeet Mishra
  \item Anne Biewald
  \item Isabelle Weindl
  \item Felicitas Beier
  \item David Chen
  \item Michael Crawford
  \item Edna Molina Bacca
  \item Ulrich Kreidenweis
  \item Amsalu W. Yalew
  \item Florian Humpenoeder 
  \item Patrick von Jeetze 
  \item Xiaoxi Wang
  \item Jan Philipp Dietrich
  \item Marcos Alves
}

}
