% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAR6SPMFig1.R
\name{readAR6SPMFig1}
\alias{readAR6SPMFig1}
\title{readAR6SPMFig1}
\usage{
readAR6SPMFig1()
}
\value{
magclass object
}
\description{
download IPCC AR6 Summary for Policymakers Fig 1, global surface temperature change
relative to 1850-1900 from observations. This dataset also includes the CMIP6 models (human and
natural forcings simulations), but are so far excluded here.
}
\examples{

\dontrun{
  readSource("AR6SPMFig1")
}

}
\seealso{
https://www.ipcc.ch/report/ar6/wg1/downloads/report/IPCC_AR6_WGI_SPM.pdf
}
\author{
Michael Crawford
}
