% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChenFoodPrices2025.R
\name{readChenFoodPrices2025}
\alias{readChenFoodPrices2025}
\title{readChenFoodPrices2025}
\usage{
readChenFoodPrices2025(subtype = "MarkupCoef")
}
\arguments{
\item{subtype}{Either "FafhCoef" for food-away-from-home share coefficients,
or "MarkupCoef" for the markup coefficients by product and consumption type}
}
\value{
magpie object with regression coefficients
}
\description{
Read regression coefficients for food price markup and food-away-from-home share
from Chen et al. 2025 Nature Food paper.
}
\examples{
\dontrun{
readSource("ChenFoodPrices2025", subtype = "MarkupCoef")
}

}
\author{
David M Chen
}
