% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEmisLucGlo.R
\name{readEmisLucGlo}
\alias{readEmisLucGlo}
\title{EmisLucGlo}
\usage{
readEmisLucGlo(subtype = NULL)
}
\arguments{
\item{subtype}{Available subtypes are:
\itemize{
\item historical:
\itemize{
\item Canadell_2007
\item Friedlingstein_2010
\item Harris_2013
\item Houghton_2012
\item RCP
}
}}
}
\value{
magpie object containing data land-use change CO2 emissions
}
\description{
Read historic land-use change CO2 emissions
}
\examples{

\dontrun{ 
a <- readSource("EmisLucGlo",subtype="Canadell_2007")
a <- readSource("EmisLucGlo",subtype="Friedlingstein_2010")
}
#'@importFrom reshape2 dcast
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Florian Humpenoeder
}
