% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOrotation.R
\name{readFAOrotation}
\alias{readFAOrotation}
\title{FAOrotation}
\usage{
readFAOrotation(subtype = "FAO")
}
\arguments{
\item{subtype}{subtype - only FAO supported}
}
\value{
magpie object containing rotation length data
}
\description{
Read Rotation lengths reported by Global planted forests thematic study: Results and analysis (2006)
}
\examples{
\dontrun{
a <- readSource("FAOrotation")
}
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Abhijeet Mishra
}
