% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGasser.R
\name{readGasser}
\alias{readGasser}
\title{Gasser}
\usage{
readGasser(subtype = "bookkeeping")
}
\arguments{
\item{subtype}{subtype}
}
\value{
magpie object containing data land-use change CARBON emissions
}
\description{
Read historic land-use change CARBON emissions
}
\examples{
\dontrun{
a <- readSource("Gasser")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra
}
