% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGISTEMP.R
\name{readGISTEMP}
\alias{readGISTEMP}
\title{readGISTEMP}
\usage{
readGISTEMP()
}
\value{
magclass object
}
\description{
read GISTEMP historical global surface temperature data from the Goddard Institute for Space Studies
}
\examples{

\dontrun{
  readSource("readGISTEMP")
}

}
\seealso{
https://data.giss.nasa.gov/gistemp/
}
\author{
Michael Crawford
}
