% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHID.R
\name{readHID}
\alias{readHID}
\title{readHID}
\usage{
readHID(subtype = NULL)
}
\arguments{
\item{subtype}{: Available subtypes are:
\itemize{
\item national_1900_2005 : National Data on AEI
\item AEI_EARTHSTAT_CP : Data in 5armin gridded resolution based on Earthstat LU scenario and maximum consistency with pasture extent
\item AEI_EARTHSTAT_IR : Data in 5armin gridded resolution based on Earthstat LU scenario and maximum consistency with  AEI from subnationals regions
\item AEI_HYDE_LOWER_CP : Data in 5armin gridded resolution based on Hyde LU scenario from the lower end of the uncertainty band and maximum consistency with pasture extent
\item AEI_HYDE_LOWER_IR : Data in 5armin gridded resolution based on Hyde LU scenario from the lower end of the uncertainty band and maximum consistency with AEI from subnational regions
\item AEI_HYDE_FINAL_CP : Data in 5armin gridded resolution based on Hyde LU scenario final version and maximum consistency with pasture extent
\item AEI_HYDE_FINAL_IR : Data in 5armin gridded resolution based on Hyde LU scenario final version and maximum consistency with AEI from subnational regions
\item AEI_HYDE_UPPER_CP : Data in 5armin gridded resolution based on Hyde LU scenario from the upper end of the uncertainty band and maximum consistency with pasture extent
\item AEI_HYDE_UPPER_IR : Data in 5armin gridded resolution based on Hyde LU scenario from the upper end of the uncertainty band and maximum consistency with AEI from subnational regions
}}
}
\value{
magpie object of the Irrigated Area  data
}
\description{
{Read Historic Irrigation Data

Read the Data from Siebert et.al on Historic Irrigation for each Country. Data contains Area equiped for Irrigation (AEI) from 1900 to 2005 in ten or five year timesteps

The Data is also available spatialy explicit with a resolution of 5 arcmin:
}
}
\details{
Further information:
\itemize{
 \item S. Siebert, M. Kummu, M. Porkka, P. D??ll, N. Ramankutty, and B.R. Scanlon. 2015.  A global data set of the extent of irrigated land from 1900 to 2005. Hydrology and Earth System Sciences. Vol. 19. p. 1521-1545. 
 \item K. K. Goldwijk, A. Beusen, G. van Drecht, and M. de Vos. 2011. The HYDE 3.1 spatially explicit database of human-induced global land-use change over the past 12,000 years. Global Ecology and Biogeography. Vol. 20. p. 73-86
 }
}
\examples{
\dontrun{ 
a <- readSource("HID", "national_1900_2005")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
