% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMPACTIrrigInvCosts.R
\name{readIMPACTIrrigInvCosts}
\alias{readIMPACTIrrigInvCosts}
\title{readIMPACTIrrigInvCosts}
\usage{
readIMPACTIrrigInvCosts()
}
\value{
magpie object containing Average annual baseline water-related investment cost on IMPACT-region level (in billion 2000 US$ per year)
}
\description{
Average annual baseline water-related investment cost data (2016-2030) read from Rosegrant et al. (2017) "Quantitative Foresight Modeling to Inform the CGIAR Research Portfolio"
}
\examples{

\dontrun{
a <- readSource("IMPACTIrrigInvCosts", convert=TRUE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Felicitas Beier
}
