% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProdPrIndex.R
\name{readProdPrIndex}
\alias{readProdPrIndex}
\title{Read FAO producer price index}
\usage{
readProdPrIndex()
}
\value{
magpie object of the world producer index from FAO
}
\description{
Read-in world producer price index as magclass object
}
\examples{

\dontrun{ a <- readSource(type="ProdPrIndex")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Mishko Stevanovic
}
