% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoilCarbonDebt.R
\name{readSoilCarbonDebt}
\alias{readSoilCarbonDebt}
\title{readSoilCarbonDebt}
\usage{
readSoilCarbonDebt()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read data from Soil Carbon Debt Paper (https://github.com/whrc/Soil-Carbon-Debt/)
}
\examples{

\dontrun{
  readSource("SoilCarbonDebt", subtype = "SOCS_2010")
}

}
\author{
Kristine Karstens
}
