% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWaterUsage.R
\name{readWaterUsage}
\alias{readWaterUsage}
\title{readWaterUsage}
\usage{
readWaterUsage(subtype = NULL)
}
\arguments{
\item{subtype}{Available subtypes are:
\itemize{
\item historical:
\itemize{
\item foley_2011
\item shiklomanov_2000
\item wada_2011
\item wisser_2008
}
\item projections
\itemize{
\item fischer_IIASA
\item hejazi_2013
\item molden_IWMI
\item seckler_IWMI
\item shiklomanov
\item aquastat_2008_12
}
}}
}
\value{
magpie object containing data on water usage
}
\description{
Read Historic and projcted Agricultural water consumption
}
\examples{
\dontrun{
a <- readSource("foley_2011")
a <- readSource("aquastat_2008_12")
}
#' @importFrom reshape2 dcast
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Stephen Wirth
}
