% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWBPoverty.R
\name{readWBPoverty}
\alias{readWBPoverty}
\title{readWBPoverty}
\usage{
readWBPoverty(subtype = "320PovertyLine")
}
\arguments{
\item{subtype}{either "Gini", or the three poverty line thresholds:
"190PovertyLine", "320PovertyLine", "550PovertyLine"}
}
\value{
magclass object
}
\description{
read World Bank poverty percentage of population under poverty line, and gini coef data
No download function from WDI API yet, as the older versions (2011PPP as opposed to 2017PPP online) are currently
most relevant for validating Soergel poverty model results and must be downloaded manually
}
\examples{
\dontrun{
readSource("readWBPoverty", subtype = "320PovertyLine")
}

}
\author{
David M Chen
}
