% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAEI.R
\name{calcValidAEI}
\alias{calcValidAEI}
\title{calcValidAEI}
\usage{
calcValidAEI(datasource = "LUH2v2")
}
\arguments{
\item{datasource}{Currently available: \code{"LUH2v2"}, \code{"HID"},
\code{"GMIA"} and \code{"Mehta2022"}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical area equipped for irrigation.
}
\author{
Stephen Wirth, Anne Biewald, Felicitas Beier
}
