% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAgEmployment.R
\name{calcValidAgEmployment}
\alias{calcValidAgEmployment}
\title{calcValidAgEmployment}
\usage{
calcValidAgEmployment(datasource = "ILO")
}
\arguments{
\item{datasource}{ILO for reporting aggregated employment in crop+livestock production, or ILO_FAO, which uses the
same aggregated employment data from ILO, but applies FAO value of production shares to disaggregated between
employment in crop and in livestock production.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
number of people employed in agriculture (crop+livestock production)
}
\examples{
\dontrun{
calcOutput("ValidAgEmployment", datasource="ILO")
}

}
\author{
Debbora Leip
}
