% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidAgGDP.R
\name{calcValidAgGDP}
\alias{calcValidAgGDP}
\title{calcValidAgGDP}
\usage{
calcValidAgGDP(datasource = "FAO")
}
\arguments{
\item{datasource}{datasource for validation. Options FAO and FAO-consum}
}
\value{
List of magpie object with results on country level, no weight, unit and description.
}
\description{
Validation for agricultural value added gdp (Million 05USD)
}
\examples{
\dontrun{
calcOutput("ValidAgGDP")
}
}
\author{
Edna J. Molina Bacca
}
