% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidBEYield.R
\name{calcValidBEYield}
\alias{calcValidBEYield}
\title{calcValidBEYield}
\usage{
calcValidBEYield()
}
\value{
List of magpie objects with results on cellular level, weight, unit, min, max, description.
}
\description{
Validation dataset for bioenergy crop yields from Li et al. (2020).
Each of the five crop types is returned as an independent entry at cellular level.
}
\examples{
\dontrun{
calcOutput("ValidBEYield", aggregate = FALSE)
}
}
\seealso{
[mrcommons::calcBEYield()]
}
\author{
Kristine Karstens
}
