% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidBII.R
\name{calcValidBII}
\alias{calcValidBII}
\title{calcValidBII}
\usage{
calcValidBII()
}
\value{
Country-level MAgPIE object
}
\description{
validation for the Biodiversity Intactness Index (BII)
}
\examples{

\dontrun{
   calcOutput("ValidBII")
}

}
\author{
Michael Crawford
}
