% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCarbon.R
\name{calcValidCarbon}
\alias{calcValidCarbon}
\title{calcValidCarbon}
\usage{
calcValidCarbon(datasource = "LPJmL4:CRU_4")
}
\arguments{
\item{datasource}{Datasources for validation data, e.g. LPJmL4:CRU_4}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for carbon pools
}
\examples{

\dontrun{ 
  calcOutput("ValidCarbon")
}

}
\author{
Kristine Karstens
}
