% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidCroparea.R
\name{calcValidCroparea}
\alias{calcValidCroparea}
\title{calcValidCroparea}
\usage{
calcValidCroparea(datasource = "FAO", detail = F)
}
\arguments{
\item{datasource}{Currently only "FAO" available}

\item{detail}{how much detail?}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical areas of individual crops. These are derived by
correcting harvested areas to match to physical cropland areas. Both these
datasets are from FAO. Output is meant to be used for model validation.
}
\author{
Ulrich Kreidenweis
}
