% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidFoodExpenditureShare.R
\name{calcValidFoodExpenditureShare}
\alias{calcValidFoodExpenditureShare}
\title{calcValidFoodExpenditureShare}
\usage{
calcValidFoodExpenditureShare(detail = FALSE)
}
\arguments{
\item{detail}{if FALSE, only major food commoditiy groups are shown.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
validation for foode expenditure share
}
\examples{

\dontrun{ 
calcOutput("ValidFoodExpenditureShare")
}

}
\author{
Benjamin Leon Bodirsky
}
