% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGlobalCarbonBudget.R
\name{calcValidGlobalCarbonBudget}
\alias{calcValidGlobalCarbonBudget}
\title{ValidGlobalCarbonBudget}
\usage{
calcValidGlobalCarbonBudget(cumulative = FALSE)
}
\arguments{
\item{cumulative}{cumulative from y2000}
}
\value{
a MAgPIE object
}
\description{
validation for total and cumulative land emissions from the Global Carbon Budget, including
all bookkeeping models
}
\examples{
\dontrun{
calcOutput("ValidGlobalCarbonBudget")
}
}
\author{
Michael Crawford
}
