% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGlobalSurfaceTemp.R
\name{calcValidGlobalSurfaceTemp}
\alias{calcValidGlobalSurfaceTemp}
\title{calcValidGlobalSurfaceTemp}
\usage{
calcValidGlobalSurfaceTemp()
}
\value{
global MAgPIE object
}
\description{
validation for the global surface temperature anomaly
}
\examples{

\dontrun{
   calcOutput("ValidGlobalSurfaceTemp", aggregate = FALSE)
}

}
\author{
Michael Crawford
}
