% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridCroparea.R
\name{calcValidGridCroparea}
\alias{calcValidGridCroparea}
\title{calcValidGridCroparea}
\usage{
calcValidGridCroparea()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Croparea by crops on 0.5 degree grid in physical area
}
\examples{

\dontrun{ 
calcOutput("ValidGridCroparea")
}

}
\seealso{
\code{\link{fullMADRATTOLPJML}}
}
\author{
Benjamin Leon Bodirsky
}
