% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGridYields.R
\name{calcValidGridYields}
\alias{calcValidGridYields}
\title{calcValidGridYields}
\usage{
calcValidGridYields()
}
\value{
List of magpie objects with results on cellular level, weight on cellular level, unit and description.
}
\description{
reports Yields on 0.5 degree grid
}
\examples{

\dontrun{ 
calcOutput("ValidGridYields")
}

}
\seealso{
\code{\link{fullMADRATtoLPJml}}
}
\author{
Benjamin Leon Bodirsky
}
