% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGS.R
\name{calcValidGS}
\alias{calcValidGS}
\title{calcValidGS}
\usage{
calcValidGS(datasource = "FAO", indicator = "relative")
}
\arguments{
\item{datasource}{Currently only  available for the "FAO" source}

\item{indicator}{type of indicator (relative or absolute)}
}
\value{
List of magpie object with growing stock
}
\description{
Returns historical FRA 2020 growing stock in million m3
}
\author{
Abhijeet Mishra
}
