% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidGS.R
\name{calcValidGS}
\alias{calcValidGS}
\title{calcValidGS}
\usage{
calcValidGS(datasource = "FRA")
}
\arguments{
\item{datasource}{Currently only  available for the "FRA" source}
}
\value{
List of magpie object with growing stock
}
\description{
Returns historical FRA 2020 growing stock in million m3
}
\author{
Abhijeet Mishra
}
