% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidHourlyLaborCosts.R
\name{calcValidHourlyLaborCosts}
\alias{calcValidHourlyLaborCosts}
\title{calcValidHourlyLaborCosts}
\usage{
calcValidHourlyLaborCosts(datasource = "ILO_completed")
}
\arguments{
\item{datasource}{So far only "ILO_completed" (based on regression between ILO hourly labor costs and GDP pc MER,
calibrated to match USDA/FAO total labor costs)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
hourly labor costs in crop+livestock production
}
\examples{
\dontrun{
calcOutput("ValidHourlyLaborCosts", datasource="ILO_completed")
}

}
\author{
Debbora Leip
}
