% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidIncome.R
\name{calcValidIncome}
\alias{calcValidIncome}
\title{calcValidIncome}
\usage{
calcValidIncome(datasource = "James")
}
\arguments{
\item{datasource}{To switch between historical values of James et al (default) or 
with future projections for different scenarios (James_OECD_Nakicenovic)}
}
\value{
list of magpie object with data and weight. Since intensive and extensive variables 
are mixed please keep the mixed_aggregation
}
\description{
Returns historical development of income and future projections of income dynamics
}
\author{
Florian Humpenoeder, Abhijeet Mishra, Kristine Karstens
}
