% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidIncome.R
\name{calcValidIncome}
\alias{calcValidIncome}
\title{calcValidIncome}
\usage{
calcValidIncome(datasource = "WDI_SSPs")
}
\arguments{
\item{datasource}{"WDI_SSPs", the only one option at the moment.}
}
\value{
list of magpie object with data and weight. Since intensive and extensive variables
are mixed please keep the mixed_aggregation
}
\description{
Returns historical development of income and future projections of income dynamics
}
\author{
Florian Humpenoeder, Abhijeet Mishra, Kristine Karstens
}
