% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidLSUdensity.R
\name{calcValidLSUdensity}
\alias{calcValidLSUdensity}
\title{calcValidLSUdensity}
\usage{
calcValidLSUdensity(luhversion = "LUH3")
}
\arguments{
\item{luhversion}{The LUH version to be used (\code{"LUH3"} or \code{"LUH2v2"})}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates historical rangelands intensity use.
}
\examples{
\dontrun{
calcOutput("ValidLSUdensity")
}

}
\author{
Marcos Alves
}
