% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPopulation.R
\name{calcValidPopulation}
\alias{calcValidPopulation}
\title{calcValidPopulation}
\usage{
calcValidPopulation(
  PopulationPast = "WDI-UN_PopDiv-MI",
  PopulationFuture = c("SSPs-UN_PopDiv-MI", "SDPs-UN_PopDiv-MI", "SSP2EU-UN_PopDiv-MI"),
  TimeFromFindSet = TRUE
)
}
\arguments{
\item{PopulationPast}{population past data source}

\item{PopulationFuture}{population future data source}

\item{TimeFromFindSet}{boolean deciding something}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical development of population and future projections of population dynamics
}
\author{
Florian Humpenoeder, Jan Philipp Dietrich, David Chen
}
