% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPopulation.R
\name{calcValidPopulation}
\alias{calcValidPopulation}
\title{calcValidPopulation}
\usage{
calcValidPopulation(
  PopulationCalib = c("past_grPEAP_grFuture", "Ariadne"),
  PopulationPast = c("WDI_completed", "Eurostat_WDI_completed"),
  PopulationFuture = c("SSP2018Update_completed_bezierOut",
    "SSP2EU_completed_bezierOut"),
  TimeFromFindSet = TRUE
)
}
\arguments{
\item{PopulationCalib}{to what should be calibrated? past, future or a transition?}

\item{PopulationPast}{population past data source}

\item{PopulationFuture}{population future data source}

\item{TimeFromFindSet}{boolean deciding something}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical development of population and future projections of population dynamics
}
\author{
Florian Humpenoeder, Jan Philipp Dietrich, David Chen
}
