% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidLandPrice.R
\name{calcValidPriceLand}
\alias{calcValidPriceLand}
\title{calcValidPriceLand}
\usage{
calcValidPriceLand(datasource = "FAO_USDA")
}
\arguments{
\item{datasource}{FAO_USDA}
}
\description{
Validates MAgPIE (crop) Land prices
}
\examples{
\dontrun{
calcOutput("ValidPriceGHG")
}

}
\author{
Edna J. Molina Bacca
}
