% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTau.R
\name{calcValidTau}
\alias{calcValidTau}
\title{calcValidTau}
\usage{
calcValidTau(datasource = "FAO2012")
}
\arguments{
\item{datasource}{Currently available: \code{"FAO2012"} (original data set)
and \code{"FAOonline"} (projection of tau values from 1995 based on recent FAO yield
projections.)}
}
\value{
tau time series
}
\description{
Returns historical land use intensity estimates (tau).
}
\author{
Jan Philipp Dietrich
}
