% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTimber.R
\name{calcValidTimber}
\alias{calcValidTimber}
\title{calcValidTimber}
\usage{
calcValidTimber(datasource = "FAO")
}
\arguments{
\item{datasource}{Currently only available for the "FAO" source}
}
\value{
List of magpie object with data and population
}
\description{
Returns historical timber demand and production from FAO in both
volumetric (Mm3/yr) and mass (mio tDM) units.
Variable names match reportTimber (Mm3) and reportTimberDemand (tDM).
}
\author{
Abhijeet Mishra, Florian Humpenoeder
}
