% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTimber.R
\name{calcValidTimber}
\alias{calcValidTimber}
\title{calcValidTimber}
\usage{
calcValidTimber(datasource = "FAO")
}
\arguments{
\item{datasource}{Currently only  available for the "FAO" source}
}
\value{
List of magpie object with data and population
}
\description{
Returns historical timber demand in mio m3 per yr
}
\author{
Abhijeet Mishra
}
