% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTotalLaborCosts.R
\name{calcValidTotalLaborCosts}
\alias{calcValidTotalLaborCosts}
\title{calcValidTotalLaborCosts}
\usage{
calcValidTotalLaborCosts(datasource = "USDA", dataVersionILO = "Aug24")
}
\arguments{
\item{datasource}{"USDA" (which uses FAO VoP and USDA cost shares) or "ILO" (which is based on ILO datasets and
calibrated to the USDA/FAO approach, but includes costs for some countries without VoP data) or "GTAP"}

\item{dataVersionILO}{"" for the oldest version, or "monthYear" (e.g. "Aug23") for a newer version}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
total labor costs in crop and livestock production
}
\examples{
\dontrun{
calcOutput("ValidTotalLaborCosts", datasource = "USDA")
}

}
\author{
Debbora Leip
}
