% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidTrade.R
\name{calcValidTrade}
\alias{calcValidTrade}
\title{calcValidTrade}
\usage{
calcValidTrade(datasource = "FAO", detail = T, nutrient = "dm")
}
\arguments{
\item{datasource}{Datasource of validation data.}

\item{detail}{if FALSE, only larger product categories are reported}

\item{nutrient}{The nutrient in which the results shall be reported.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the validation data for trade of agricultural products
}
\examples{

\dontrun{ 
calcOutput("ValidTrade")
}

}
\seealso{
\code{\link{calcFAOmassbalance}},
\code{\link{calcValidDemand}}
}
\author{
Benjamin Leon Bodirsky, Xiaoxi Wang
}
