% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidWageDevelopment.R
\name{calcValidWageDevelopment}
\alias{calcValidWageDevelopment}
\title{calcValidWageDevelopment}
\usage{
calcValidWageDevelopment(
  datasource = "ILO_completed",
  baseYear = 2000,
  dataVersionILO = "Aug24"
)
}
\arguments{
\item{datasource}{Available datasources are:
\itemize{
\item ILO_raw : based on ILO hourly labor costs data
\item ILO_completed : based on ILO hourly labor costs data completed with a regression with GDP pc MER
\item USDA_FA0_raw : based on USDA/FAO hourly labor costs data
\item USDA_FA0_completed : based on USDA/FAO hourly labor costs data completed with a regression with GDP pc MER
}}

\item{baseYear}{year relative to which the wage development should be calculated}

\item{dataVersionILO}{"" for the oldest version, or "monthYear" (e.g. "Aug23") for a newer version}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
wage index: hourly labor costs in crop+livestock production relative to a baseyear
}
\examples{
\dontrun{
calcOutput("ValidWageDevelopment", datasource = "ILO_completed")
}

}
\author{
Debbora Leip
}
