% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidWaterUsage.R
\name{calcValidWaterUsage}
\alias{calcValidWaterUsage}
\title{calcValidWaterUsage}
\usage{
calcValidWaterUsage(datasource = "shiklomanov_2000")
}
\arguments{
\item{datasource}{Currently available: 
\itemize{
\item historical:
\itemize{
\item \code{"foley_2011"}
\item \code{"shiklomanov_2000"}
\item \code{"wada_2011"}
\item \code{"wisser_2008"}
\item \code{"cwatm_ipsl-cm5a-lr"}
\item \code{"cwatm_gfdl-esm2m"}
\item \code{"cwatm_miroc5"}
\item \code{"cwatm_hadgem2-es"}
\item \code{"lpjml_ipsl-cm5a-lr"}    
\item \code{"lpjml_gfdl-esm2m"}
\item \code{"lpjml_miroc5"}
\item \code{"lpjml_hadgem2-es"}
\item \code{"h08_ipsl-cm5a-lr"}    
\item \code{"h08_gfdl-esm2m"}
\item \code{"h08_miroc5"}
\item \code{"h08_hadgem2-es"}
\item \code{"matsiro_ipsl-cm5a-lr"}
\item \code{"matsiro_gfdl-esm2m"}    
\item \code{"matsiro_miroc5"}
\item \code{"matsiro_hadgem2-es"}
\item \code{"mpi-hm_ipsl-cm5a-lr"}
\item \code{"mpi-hm_gfdl-esm2m"}
\item \code{"mpi-hm_miroc5"}    
\item \code{"pcr-globwb_ipsl-cm5a-lr"}
\item \code{"pcr-globwb_gfdl-esm2m"}
\item \code{"pcr-globwb_miroc5"}
\item \code{"pcr-globwb_hadgem2-es"}
}
\item projections:
\itemize{
\item \code{"fischer_IIASA"}
\item \code{"hejazi_2013"}
 \item \code{molden_IWMI}
 \item \code{seckler_IWMI}
 \item \code{shiklomanov}
}
}}
}
\value{
list of magpie object with data and weight
}
\description{
Returns historical and projected water usage from different sources
}
\author{
Stephen Wirth, Anne Biewald, Felicitas Beier
}
