% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidYield.R
\name{calcValidYield}
\alias{calcValidYield}
\title{calcValidYield}
\usage{
calcValidYield(datasource = "FAO", future = NULL)
}
\arguments{
\item{datasource}{Specify which datasource needs to be used.
Currently only "FAO" and "calibratedLPJmL" is available.}

\item{future}{if NULL no future values are returned (default).
specify climate scenario (gcm:rcp), if future is needed}
}
\value{
List of magpie objects with results on country level,
        weight on country level, unit,
        Max.&Min. values alongwith description.
}
\description{
Calculates a dataset of agricultural production out of the combined
data of calcFAOharmonized(). Covers dry matter (DM) production.
Also returns areas of individual crops from FAOSTAT.
Total area can be lower or higher than arable land because of multicropping or fallow land.
Returns yield as calculated from area area and production.
}
\examples{
\dontrun{
calcOutput("ValidYield")
}
}
\seealso{
\code{\link{calcFAOmassbalance}},
\code{\link{calcCroparea}}
}
\author{
Abhijeet Mishra, Isabelle Weindl
}
