% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadBII.R
\name{downloadBII}
\alias{downloadBII}
\title{downloadBII}
\usage{
downloadBII()
}
\value{
Metadata on downloaded BII data
}
\description{
Download historical and projected BII dataset from Phillips et al. 2021
(https://doi.org/10.5519/he1eqmg1).
}
\examples{

\dontrun{
  downloadSource("BII")
}

}
\seealso{
https://data.nhm.ac.uk/dataset/bii-bte
}
\author{
Michael Crawford
}
