% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMADRATtoLPJml.R
\name{fullMADRATtoLPJml}
\alias{fullMADRATtoLPJml}
\title{fullMADRATtoLPJml}
\usage{
fullMADRATtoLPJml()
}
\value{
No return; the function writes a number of netcdf files into the output folder of MADRAT
}
\description{
creates historical outputs to be used by LPJml or other models that require gridded inputs
}
\examples{

\dontrun{ 
fullMADRATtoLPJml()
}

}
\seealso{
\code{\link{fullVALIDATION}}
}
\author{
Benjamin Leon Bodirsky
}
