% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalCarbonBudget.R
\name{readGlobalCarbonBudget}
\alias{readGlobalCarbonBudget}
\title{readGlobalCarbonBudget}
\usage{
readGlobalCarbonBudget()
}
\value{
a magpie object in Mt CO2 per year
}
\description{
read the Global Carbon Budget, selecting the models GCB, BLUE, H&C2023, OSCAR and their sub-components
Net, Deforestation, Forest regrowth, Other transitions, Wood harvest and other forest management
}
\examples{
\dontrun{
readSource("GlobalCarbonBudget")
}
}
\author{
Michael Crawford
}
